--- gismoclouddeploy/services/client/utils/k8s.py
@@ -1,13 +1,0 @@
-from kubernetes import client, config
-
-def get_k8s_pod_name(container_name):
-    config.load_kube_config()
-    v1 = client.CoreV1Api()
-    print("Listing pods with their IPs:")
-    ret = v1.list_pod_for_all_namespaces(watch=False)
-    for i in ret.items:
-        # print("%s\t%s\t%s" % (i.status.pod_ip, i.metadata.namespace, i.metadata.name))
-        podname = i.metadata.name.split("-")[0]
-        if podname == container_name:
-            print(f"podname: {i.metadata.name}")
-            return i.metadata.name
